requirements.txt
# AWS SDK
boto3==1.28.38

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20

# Data Processing
pandas==2.0.3
numpy==1.24.4

# Web Requests
requests==2.31.0
beautifulsoup4==4.12.2

# Environment Configuration
python-dotenv==1.0.0

# Testing and Development
pytest==7.4.0
faker==19.3.1

# API Integration
openai==0.28.0
google-generativeai==0.2.1

# Utilities
uuid==1.30
